Delete and add manually working!!!

+++ functions_emp.py
@@ -4,6 +4,7 @@ import tkinter as tk
 from PIL import ImageTk, Image
 from tkinter import simpledialog
 from tkinter import messagebox
+from tkinter import filedialog
 
 project_folder = Path("D:\\Python\\")
 open_employee = project_folder / 'employees.csv'
@@ -15,23 +16,27 @@ def add_employee():
     new_id = simpledialog.askinteger("input", "Please enter ID")
     if new_id is None:
         messagebox.showerror("You must enter only ID number")
-    if new_id in employees.id.values:
+    if new_id in employees.employee_id.values:
         messagebox.showerror("ID error", "This ID already exists!")
     new_name = simpledialog.askstring("input name", "Please enter the Name")
-    if new is None:
+    if new_name is None:
         messagebox.showerror("You must enter only letters")
+        add_employee()
+    if not new_name.isalpha():
+        messagebox.showerror("You must enter only letters")
+        add_employee()
     new_phone = simpledialog.askinteger("input phone", "Please enter Phone number")        
     if new_phone is None:
         messagebox.showerror("You must enter only numbers")        
     new_age = simpledialog.askinteger("input age", "Please enter the age")
     if new_age is None:
         messagebox.showerror("You must enter only numbers")
-    if len(new_age) > 2:
+    if len(str(new_age)) > 2:
         messagebox.showerror("You must enter only 2 numbers for age") 
     
     # Pass the info to DataFrame’s constructor to create a dataframe object
     info_df = pd.DataFrame(
-        {'employee_id': new_id, 'name': new_name, 'phone': new_phone, 'age': new_age})
+        {'employee_id': [new_id], 'name': [new_name], 'phone': [new_phone], 'age': [new_age]})
     print(info_df)
     # appends the add_file to the employees file w new index
     employees = pd.concat([employees, info_df], ignore_index=True)
@@ -39,20 +44,20 @@ def add_employee():
     print(employees)
     employees.to_csv(open_employee, index=False)
     messagebox.showinfo("New Employee", "New employee added!") 
+
+
+
 # +++++++++++++++++ Add from file
 
-
 def add_e_fromFile():
-    try:
-        new = input("Please enter the full path of the file:\n")
-    # checking for errors:
-    except ValueError:
-        print("You must enter the FULL path")
+    new = filedialog.askopenfilename()
+    '''
     employees = pd.read_csv(open_employee, index_col=False,
                             skipinitialspace=True, skip_blank_lines=True)
     # The file Must be w\o any blank lines!!
     add_file = pd.read_csv(new, index_col=False, skipinitialspace=True,
                            skip_blank_lines=True)  # reads the file to be added
+    '''
     # Check if all the data of all employees is supplied
     # It returns True if all elements within a series or along a Dataframe axis are non-zero, not-empty or not-False.טע==
     print(add_file.isna().any().any())
@@ -86,20 +91,28 @@ def delete_employee():
 
     emp = employees.drop_duplicates()  # deletes any duplicates in file
     print(emp)
-    
-    try:
-        new = input("Please enter the name of the employee to delete:\n")
-    # checking for errors:
-    except ValueError:
-        print("You must enter only a name")
-    if new not in emp['name']:
-        print('No such employee in our file')
-        # takes up all the names except the input(new), thereby dropping the row with input name
-    emp = emp[employees.name != new]
-    print(emp)
-    emp.to_csv(open_employee, index=False)
-    # The working path HAVE to be the Final_Project Path or it won't write to the right file!!!
-
+    new_name = simpledialog.askstring("input", "Please enter the name to DELETE")
+    if new_name is None:
+        return
+    elif not new_name.isalpha():
+        messagebox.showerror("Error", "You must enter a name")
+        return
+    if new_name not in employees.name.values:
+        messagebox.showerror("Name error", "This NAME doesn't exist!")
+        return
+    else:
+        confirmtxt = "Are you sure you want to delete " + new_name +"?"
+        if messagebox.askokcancel("Delete", confirmtxt, default="ok", icon="warning"):
+             
+            # takes up all the names except the input(new), thereby dropping the row with input name
+            emp = emp[employees.name != new_name]
+            print(emp)
+            emp.to_csv(open_employee, index=False)
+            # The working path HAVE to be the Final_Project Path or it won't write to the right file!!!
+            messagebox.showinfo("Delete", "Employee Deleted")
+        else:
+            messagebox.showinfo("Delete", "Delete employee - CANCELED!")
+# ------------------------Delete employee from file        
 # Delete employees from a file
 def delete_fromFile():
     try:
